indexer/src/pyproject.toml
[tool.poetry]
name = "bingo-indexer"
version = "0.1.0"
description = ""
authors = ["Lucas <lucaspar@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.6"
nltk = "^3.4"
bs4 = "^0.0.1"
boto3 = "^1.10"
pprint = "^0.1.0"
python-dotenv = "^0.10.3"
pymongo = "^3.9"
colorlog = "^4.0"
yapf = "^0.29.0"
rope = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
